requirements.txt
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
customtkinter==5.2.1
darkdetect==0.8.0
docutils==0.20.1
flatbuffers==23.5.26
humanfriendly==10.0
idna==3.6
imageio==2.33.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
Kivy==2.2.1
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.6.0
Kivy-Garden==0.1.5
lazy_loader==0.3
llvmlite==0.41.1
mpmath==1.3.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
onnxruntime==1.16.3
opencv-python-headless==4.8.1.78
packaging==23.2
Pillow==10.1.0
platformdirs==4.1.0
pooch==1.8.0
protobuf==4.25.1
Pygments==2.17.2
PyMatting==1.1.12
pypiwin32==223
pyreadline3==3.4.1
pywin32==306
referencing==0.32.0
rembg==2.0.53
requests==2.31.0
rpds-py==0.16.2
scikit-image==0.22.0
scipy==1.11.4
sd-parsers==0.3rc2
sympy==1.12
termcolor==2.4.0
tifffile==2023.12.9
tqdm==4.66.1
urllib3==2.1.0
